// types/userData.d.ts
export interface UserData {
  id: number;
  old_id: number;
  emp_no: number;
  emp_id: string;
  name: string;
  person_id: number;
  division_id: number;
  approval_level: number;
  role_id: number;
  email: string;
  phone_number: string;
  signature: number;
  images: number;
  profile_images: string;
  firebase_token: string;
  inactive: number;
  created_at: string;
  updated_at: string;
}